backend/requirements.txt
# Requirements
deepagents
langgraph
langchain-core
fastapi
uvicorn
pydantic
python-dotenv
httpx
markdownify
tavily-python
langchain
langchain-google-genai
langchain-cerebras
langfuse
